requirements.txt
alembic==1.0.10
astroid==2.3.1
Babel==2.7.0
blinker==1.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
dominate==2.4.0
elasticsearch==7.0.5
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.1
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.0.10
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.4.2
python-dateutil==2.8.0
python-dotenv==0.10.2
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.4.0
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
